+++ R/as_design_tibble.R
@@ -55,7 +55,7 @@ as_design_tibble.rigma_get_file_styles <- function(
 
   rigma_resp %>%
     chuck("meta", "styles") %>%
-    map_dfr(~{.x %>% as_tibble()}) %>%
+    map_dfr(as_tibble) %>%
     nest(user_data = "user") %>%
     mutate(
       created_at = as_datetime(.data$created_at),