+++ R/get_component.R
@@ -7,9 +7,6 @@
 #'
 #' @returns S3 object of class `rigma_get_component`.
 #'
-#' @importFrom httr2 request req_url_path_append req_headers req_user_agent
-#' req_perform resp_body_json req_url_query
-#'
 #' @importFrom checkmate assert_string
 #'
 #' @examples
@@ -22,11 +19,12 @@
 get_component <- function(key) {
   assert_string(key)
 
-  resp <- request("https://api.figma.com/v1/components/") %>%
-    req_url_path_append(key) %>%
-    req_rigma_agent %>%
-    req_perform() %>%
-    resp_body_json()
+  resp <- request_figma() %>%
+    req_figma_template(
+      "component",
+      key = key,
+      .perform = TRUE
+    )
 
   structure(
     resp,