+++ R/get_team_component_sets.R
@@ -20,9 +20,6 @@
 #' @returns S3 object of class `rigma_get_team_component_sets`. Components are
 #' stored in the `meta` field.
 #'
-#' @importFrom httr2 request req_url_path_append req_headers req_user_agent
-#' req_perform resp_body_json req_url_query
-#'
 #' @importFrom checkmate assert_string assert_number
 #'
 #' @examplesIf Sys.getenv("FIGMA_ACCESS_TOKEN") != ""
@@ -43,24 +40,16 @@ get_team_component_sets <- function(
   assert_number(after, null.ok = TRUE)
   assert_number(before, null.ok = TRUE)
 
-  params <- list(
-    page_size = page_size,
-    after = after,
-    before = before
-  )
-
-  resp <- request("https://api.figma.com/v1/teams/") %>%
-    req_url_path_append(team_id) %>%
-    req_url_path_append("component_sets") %>%
-    req_url_query(!!!params) %>%
-    req_error(body = function(resp) {
-      resp %>%
-        resp_body_json() %>%
-        chuck("message")
-    }) %>%
-    req_rigma_agent %>%
-    req_perform() %>%
-    resp_body_json()
+  resp <- request_figma() %>%
+    req_figma_template(
+      "team component sets",
+      team_id = team_id
+    ) %>%
+    req_figma_query(
+      page_size = page_size,
+      after = after,
+      before = before
+    )
 
   structure(
     list(