+++ R/req_rigma_agent.R
@@ -4,8 +4,12 @@
 #' performing the request.
 #'
 #' @param req httr2_request object
-#'
-#' @param user_agent string with user agent
+#' @param token Figma API access token. If `NULL`, replace with environment
+#'   variable named "FIGMA_ACCESS_TOKEN" or an alternate name set using the
+#'   variable "Rigma.figma_token".
+#' @param user_agent string with user agent. Defaults to "Rigma
+#'   https://github.com/AleKoure/Rigma"
+#' @keywords internal
 #'
 #' @importFrom httr2 req_headers req_user_agent req_retry resp_status
 #'
@@ -16,10 +20,15 @@
 #'
 req_rigma_agent <- function(
     req,
+    token = NULL,
     user_agent = "Rigma https://github.com/AleKoure/Rigma"
 ) {
   assert_class(req, "httr2_request")
-  figma_token <- Sys.getenv("FIGMA_ACCESS_TOKEN")
+  figma_token <- token %||%
+    Sys.getenv(
+      getOption("Rigma.figma_token", "FIGMA_ACCESS_TOKEN")
+    )
+
   assert_string(figma_token, min.chars = 1)
 
   req %>%